[C#-4]Adding striped solution
